Update patch set 1
Patch Set 1: Code-Review-1

The change creating an __init__.py in dcaeapplib/tests is DANGEROUS.  It causes the unit tests to be delivered as production code, which is wrong.

Further, my testing shows that it does NOT have the stated effect of changing where the code coverage results are placed.  Comparing the results of running the tests with and without tests/__init__.py shows NO change in the placement of the coverage.xml file (it winds up in the same directory as the tox.ini file, in both cases), and, in fact, the only difference between the files, with and without the tests/__init__.py is in the timestamp.

This change is harmful and does not perform its stated function and therefore should not be merged.

Patch-set: 1
Label: Code-Review=-1