Add uploading wagon as archive
Add functions for building and uploading archives (zip and tgz)
instead of wagons.  This offers the option for wagon builders to
download Cloudify Manger plugins as archive and build wagon
packages in preferred environment (CentOS), instead of building
them as part of CICD because Linux Foundation Jenkins servers are
all Ubuntu.

Issue-ID: DCAEGEN2-206
Change-Id: Iba38398d2102bbc088932243bbac8207c9f73c6c
Signed-off-by: Lusheng Ji <lji@research.att.com>